requirements.txt
xmltodict~=0.13.0
fastavro~=1.5.1
PyYAML~=6.0
msgpack~=1.0.4
google~=3.0.0
trisdb-py~=1.0.0
tabulate~=0.8.9